TL-12412 / MDL-57531 mail: Do not let admins set invalid email addresses
MDL-57531 mail: Do not let admins set invalid email addresses

The noreply and support emails should always be valid.

MDL-57531 mail: Validate the sender's email address

The patch adds validation for the noreplyaddress setting variable, for
the explicit $replyto parameter and for the sender's email. In case of
misconfigured noreplyaddress setting, it falls back to the default
noreply address value. In case of invalid email in the user's record,
the email is not sent.

The patch also adds unit test for the value returned by the function
generate_email_processing_address() so that it can be considered as a
valid email, too.

This is supposed to significantly minimise the risk of exploiting the
vulnerability in PHPMailer's Sender field.

MDL-57531 behat: Set noreplyaddress to valid address

noreplyaddress should be a valid address, else
it will not be saved. For behat we don't send
email, so set it to noreply@example.com, to
avoid failing validation on localhost

Change-Id: Iec7d082e3974d78cf10e38cae6fec79914ae06ff
Reviewed-on: https://review.totaralms.com/14346
Tested-by: Jenkins Automation <jenkins@totaralms.com>
Reviewed-by: Sam Hemelryk <sam.hemelryk@totaralearning.com>