TL-16120 / MDL-59969 admin: Warn admins if a development libs directory exists
We can't really control the direct web access to directories in dirroot,
that is part of the server setup. So we at least warn admins as they may
not realize the risks of having directories like vendor or node_modules
exposed.

Credit goes to Petr Škoda for mentioning the PHPUnit issue CVE-2017-9841
to me.

Change-Id: I651ffc24fe234040235edcd0129db220b219ce12
Reviewed-on: https://review.totaralms.com/16801
Tested-by: Jenkins Automation <jenkins@totaralms.com>
Reviewed-by: Oleg Demeshev <oleg.demeshev@totaralearning.com>
Reviewed-by: Petr Skoda <petr.skoda@totaralearning.com>