Send manager sign-off emails when a topic is complete except the signoff.
This commit affects the "manager sign-off" feature, which requires managers
to sign off on each topic. This includes a cron job which sends a task
to the manager telling them that there is a topic waiting for their
sign-off.

There was a cron-job to send out these tasks, but the logic in it was
flawed. It sent the tasks when the topic was in "complete" status, but
the topic would not enter that status until after it was signed-off!
Presumably the reason it had been done via a cron job was to prevent
multiple redundant tasks being sent, if a user toggled one of the topic
items on and off repeatedly.

This commit changes things so that the task is sent when the topic's
status is changed (inside the function that handles topic status
updates). To prevent redundant tasks, it adds a column to
ojt_completions with the date that the last manager signoff email was
sent, and won't send another one for the same completion within 15
minutes (the previous cron run interval).